fixes uploading duplicate images

+++ app/models/media_object.rb
@@ -37,7 +37,9 @@ class MediaObject < ActiveRecord::Base
   end
 
   def make_md5
-    self.md5 = Digest::MD5.file(file_name).hexdigest
+    if File.exist? file_name
+      self.md5 = Digest::MD5.file(file_name).hexdigest
+    end
   end
 
   def unique_md5
@@ -154,13 +156,13 @@ class MediaObject < ActiveRecord::Base
   end
 
   # Creates a deep clone of a mediaobject, including copying the underlying file
+  # Doesn't save the media object, owner and user ID must still be updated
   def cloneMedia
     nmo = dup
     nmo.store_key = nil
     nmo.check_store!
     FileUtils.cp(file_name, nmo.file_name)
     nmo.add_tn
-    nmo.save!
     nmo
   end
 
@@ -169,17 +171,20 @@ class MediaObject < ActiveRecord::Base
     unless owner_id.nil?
       self.user_id = owner_id
     end
+
     self.media_type = 'image'
     self.name = 'Uploaded Image ' + SecureRandom.hex[0...5] + "#{file_type}"
     self.file = name.split('.')[0] + SecureRandom.hex + '.' + name.split('.')[1]
     sanitize_media
     self.store_key = nil
     self.check_store!
+
     File.open("#{file_name}", 'wb+') do |f|
       f.write image_data
       f.chmod(0644)
     end
     add_tn
+
     self.save!
   end
 
@@ -195,9 +200,14 @@ class MediaObject < ActiveRecord::Base
         else
           file_type = '.jpg'
         end
-        summernote_mo = MediaObject.new
-        summernote_mo.summernote_image(data, file_type, type, item_id, owner_id)
-        picture['src'] = summernote_mo.src
+        begin
+          summernote_mo = MediaObject.new
+          summernote_mo.summernote_image(data, file_type, type, item_id, owner_id)
+          picture['src'] = summernote_mo.src
+        rescue
+          summernote_mo = MediaObject.where("#{type} = ? AND md5 = ?", item_id, summernote_mo.md5).first
+          picture['src'] = summernote_mo.src
+        end
       end
     end
     text.search('iframe').each do |video|
@@ -219,8 +229,10 @@ class MediaObject < ActiveRecord::Base
   private
 
   def remove_data!
-    File.delete(file_name)
-    File.delete(tn_file_name)
+    unless store_key.nil?
+      File.delete(file_name)
+      File.delete(tn_file_name)
+    end
   rescue Errno::ENOENT
     # whatever
   end
